+++ src/main/java/com/monthlyib/server/openapi/user/dto/GoogleLoginRequest.java
@@ -1,0 +1,17 @@
+package com.monthlyib.server.openapi.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class GoogleLoginRequest {
+
+    private String authorizationCode;
+
+    private String redirectUri;
+}